git: don't die when current user is not in /etc/passwd
This patch is necessary in Docker containers which run as
a non-root user and that non-root user is not in /etc/passwd.
A common situation is a 'build tools' container which is called
for compiling/building software artifacts and therefore should
use the current user's UID/GID. Otherwise new files are created
with the user/group set to 'root', which is problematic for
further processing of the files.

The patch returns the information which is present in /etc/passwd,
but replaces most of it with 'dummy' info. Only the current UID
and GID are set correctly with getuid() and getgid() calls.